requirements.txt
langchain
langchain-core
streamlit
langchain_community
python-dotenv
fastapi
uvicorn
langserve
sse_starlette